+++ packages/components/modules/__shared__/web/SocialInput/index.tsx
@@ -7,7 +7,7 @@ import { SocialTextField as DefaultSocialTextField } from '@baseapp-frontend/des
 import { SOCIAL_UPSERT_FORM } from '../../common'
 import DefaultSocialUpsertActions from '../SocialUpsertActions'
 import DefaultSubmitActions from './SubmitActions'
-import { Form } from './styled'
+import { Form as DefaultForm } from './styled'
 import { SocialInputProps } from './types'
 
 /**
@@ -63,11 +63,13 @@ const SocialInput = forwardRef<HTMLInputElement, SocialInputProps>(
     {
       placeholder = 'Message...',
       autoFocusInput,
+      mode,
       SocialTextField = DefaultSocialTextField,
       SocialTextFieldProps = {},
       SocialUpsertActions = DefaultSocialUpsertActions,
       SubmitActions = DefaultSubmitActions,
       SubmitActionsProps = {},
+      Form = DefaultForm,
       onKeyDown,
       formId = 'text-field-form',
       submit,
@@ -105,6 +107,7 @@ const SocialInput = forwardRef<HTMLInputElement, SocialInputProps>(
           onKeyDown={handleKeyDown}
           placeholder={placeholder}
           autoFocus={autoFocusInput}
+          mode={mode}
           isReply={isReply}
           replyTargetName={replyTargetName}
           onCancelReply={onCancelReply}