+++ packages/components/modules/__shared__/web/SocialInput/types.ts
@@ -1,6 +1,9 @@
-import { FC, KeyboardEvent } from 'react'
+import { ComponentPropsWithoutRef, ComponentType, FC, KeyboardEvent } from 'react'
 
-import { SocialTextFieldProps } from '@baseapp-frontend/design-system/components/web/inputs'
+import {
+  SocialTextFieldMode,
+  SocialTextFieldProps,
+} from '@baseapp-frontend/design-system/components/web/inputs'
 
 import { UseFormReturn } from 'react-hook-form'
 
@@ -10,11 +13,13 @@ import { SubmitActionsProps } from './SubmitActions/types'
 export interface SocialInputProps {
   placeholder?: string
   autoFocusInput?: boolean
+  mode?: SocialTextFieldMode
   SocialTextField?: FC<SocialTextFieldProps>
   SocialTextFieldProps?: Partial<SocialTextFieldProps>
   SocialUpsertActions?: FC
   SubmitActions?: FC<SubmitActionsProps>
   SubmitActionsProps?: Partial<SubmitActionsProps>
+  Form?: ComponentType<ComponentPropsWithoutRef<'form'>>
   onKeyDown?: (event: KeyboardEvent<HTMLDivElement>, onSubmit: VoidFunction) => void
   formId?: string
   submit: (data: SocialUpsertForm) => void