+++ packages/components/modules/comments/web/CommentItem/CommentReactionButton/index.tsx
@@ -11,8 +11,8 @@ import { Typography } from '@mui/material'
 import { ReactionButton } from '../../../../__shared__/web'
 import { CommentReactionButtonProps } from './types'
 
-const CommentReactionButton: FC<CommentReactionButtonProps> = ({ target: targetRef }) => (
-  <ReactionButton target={targetRef} reactionType="LIKE">
+const CommentReactionButton: FC<CommentReactionButtonProps> = ({ target: targetRef, onClick }) => (
+  <ReactionButton onClick={onClick} target={targetRef} reactionType="LIKE">
     {({ handleReaction, target }) => (
       <div className="grid grid-cols-[repeat(2,minmax(20px,max-content))] gap-1">
         <IconButton onClick={handleReaction} aria-label={`react to comment ${target.id}`}>