+++ packages/components/modules/comments/web/CommentItem/CommentsReplies/types.ts
@@ -3,7 +3,7 @@ import type { FC } from 'react'
 import type { VirtuosoProps } from 'react-virtuoso'
 
 import { CommentsList_comments$key } from '../../../../../__generated__/CommentsList_comments.graphql'
-import type { CommentItemProps } from '../types'
+import type { CommentItemProps, CustomizableCommentItemProps } from '../types'
 
 export interface CommentsRepliesProps {
   target: CommentsList_comments$key
@@ -12,4 +12,5 @@ export interface CommentsRepliesProps {
   onReplyClick?: () => void
   CommentItem: FC<CommentItemProps>
   VirtuosoProps?: Partial<VirtuosoProps<any, any>>
+  CommentItemProps?: CustomizableCommentItemProps
 }