+++ packages/components/modules/comments/web/index.ts
@@ -12,6 +12,9 @@ export type * from './CommentUpdate/types'
 export { default as Comments } from './Comments'
 export type * from './Comments/types'
 
+export { default as BaseComments } from './BaseComments'
+export type * from './BaseComments/types'
+
 export { default as CommentsList } from './CommentsList'
 export type * from './CommentsList/types'