+++ packages/components/modules/messages/common/graphql/queries/GroupDetailsQuery.ts
@@ -6,6 +6,7 @@ export const GroupDetailsQuery = graphql`
       id
       participantsCount
       isArchived
+      participantIds
       ...RoomTitleFragment
       ...MembersListFragment
     }