+++ packages/components/modules/notifications/web/NotificationsList/NotificationItem/ReactionCreated/index.tsx
@@ -2,29 +2,43 @@ import { FC } from 'react'
 
 import { formatRelativeTime } from '@baseapp-frontend/utils'
 
-import Notification from '../Notification'
-import NotificationContent from '../Notification/NotificationContent'
-import { GenericItemProps } from '../types'
+import DefaultNotificationAvatar from '../Notification/NotificationAvatar'
+import DefaultNotificationHeader from '../Notification/NotificationHeader'
+import {
+  NotificationContent as DefaultNotificationContent,
+  NotificationRoot as DefaultNotificationRoot,
+} from '../styled'
+import { ReactionCreatedProps } from './types'
 
-const ReactionCreated: FC<GenericItemProps> = ({ notification }) => {
+const ReactionCreated: FC<ReactionCreatedProps> = ({
+  notification,
+  NotificationRoot = DefaultNotificationRoot,
+  NotificationContent = DefaultNotificationContent,
+  NotificationAvatar = DefaultNotificationAvatar,
+  NotificationAvatarProps = {},
+  NotificationHeader = DefaultNotificationHeader,
+  NotificationHeaderProps = {},
+}) => {
   // eslint-disable-next-line no-underscore-dangle
-  const message = `liked your ${notification.target?.__typename.toLowerCase()}`
+  const message = `liked your ${notification.target?.__typename?.toLowerCase?.() ?? ''}`
 
   return (
-    <Notification.Root>
-      <Notification.Avatar
+    <NotificationRoot>
+      <NotificationAvatar
         actorAvatar={notification.actor?.image?.url}
         actorName={notification.actor?.name}
+        {...NotificationAvatarProps}
       />
       <NotificationContent>
-        <NotificationContent.Header
+        <NotificationHeader
           message={message}
           timestamp={formatRelativeTime(notification.timestamp)}
           actorName={notification.actor?.name}
           unread={notification.unread}
+          {...NotificationHeaderProps}
         />
       </NotificationContent>
-    </Notification.Root>
+    </NotificationRoot>
   )
 }