+++ packages/components/modules/notifications/web/index.ts
@@ -9,8 +9,6 @@ export type * from './NotificationsList/types'
 export { default as NotificationItem } from './NotificationsList/NotificationItem'
 export type * from './NotificationsList/NotificationItem/types'
 
-export { default as Notification } from './NotificationsList/NotificationItem/Notification'
-
 export { default as NotificationItemRenderer } from './NotificationsList/NotificationItem/NotificationItemRenderer'
 export type * from './NotificationsList/NotificationItem/NotificationItemRenderer/types'
 
@@ -19,3 +17,18 @@ export { default as CommentCreated } from './NotificationsList/NotificationItem/
 export { default as CommentReply } from './NotificationsList/NotificationItem/CommentReply'
 
 export { default as ReactionCreated } from './NotificationsList/NotificationItem/ReactionCreated'
+
+export { default as NotificationAvatar } from './NotificationsList/NotificationItem/Notification/NotificationAvatar'
+export type * from './NotificationsList/NotificationItem/Notification/NotificationAvatar/types'
+
+export { default as NotificationHeader } from './NotificationsList/NotificationItem/Notification/NotificationHeader'
+export type * from './NotificationsList/NotificationItem/Notification/NotificationHeader/types'
+
+export { default as NotificationBody } from './NotificationsList/NotificationItem/Notification/NotificationBody'
+export type * from './NotificationsList/NotificationItem/Notification/NotificationBody/types'
+
+export { default as MarkAllAsReadButton } from './NotificationsList/MarkAllAsReadButton'
+export type * from './NotificationsList/MarkAllAsReadButton/types'
+
+export { default as EmptyState } from './NotificationsList/EmptyState'
+export type * from './NotificationsList/EmptyState/types'