+++ packages/components/modules/profiles/common/graphql/queries/Profile.ts
@@ -5,7 +5,6 @@ export const ProfileByIdQuery = graphql`
     profile(id: $id) {
       ...CheckMountPermissionWrapperFragment @arguments(perm: $perm)
       ...ProfileComponentFragment
-      ...FollowToggleUpdatableFragment
     }
   }
 `