+++ packages/components/modules/profiles/common/utils/getImageUrl/index.ts
@@ -1,0 +1,6 @@
+// TODO: maybe move this to  @baseapp-frontend/utils
+export const getImageUrl = (image?: string | File | Blob | MediaSource | null) => {
+  if (typeof image === 'string') return image
+  if (image instanceof Blob) return URL.createObjectURL(image)
+  return ''
+}