+++ packages/design-system/components/web/avatars/AvatarWithPlaceholder/types.ts
@@ -3,5 +3,7 @@ import { AvatarProps } from '@mui/material'
 export interface AvatarWithPlaceholderProps extends AvatarProps {
   width?: number
   height?: number
+  borderStyle?: 'solid' | 'dashed' | 'dotted'
+  borderWidth?: string
   showDeletedUser?: boolean
 }